Second

+++ Application/DTOs/GetHotelDto.cs
@@ -11,7 +11,8 @@ namespace Application.DTOs
         public int Id { get; set; }
         public string Name { get; set; } = null!;
         public int Rating { get; set; }
-        public int Price { get; set; } 
+        
+        public decimal Price { get; set; } 
 
     }
 }
